Uniformed e2e cruise test + small fixes
