+++ test/e2e/cruise/conflictremotenamespace/conflict_remote_namespace_test.go
@@ -39,10 +39,20 @@ import (
 const (
 	// clustersRequired is the number of clusters required in this E2E test.
 	clustersRequired = 3
-	// testNamespaceName is the name of the test namespace for this test.
-	testNamespaceName = "test-namespace-conflict"
 	// testName is the name of this E2E test.
-	testName = "E2E_CONFLICT_CREATION"
+	testName = "CONFLICT_REMOTE_NAMESPACE"
+)
+
+var (
+	ctx           = context.Background()
+	testContext   = tester.GetTester(ctx)
+	interval      = 1 * time.Second
+	timeout       = 10 * time.Second
+	localIndex    = 0
+	namespaceName = util.GetNameNamespaceTest(testName)
+
+	// index of the cluster on which a remote namespace with the same name already exists.
+	remoteIndex = 2
 )
 
 func TestE2E(t *testing.T) {
@@ -52,16 +62,6 @@ func TestE2E(t *testing.T) {
 }
 
 var _ = Describe("Liqo E2E", func() {
-	var (
-		ctx         = context.Background()
-		testContext = tester.GetTester(ctx)
-		interval    = 1 * time.Second
-		timeout     = 10 * time.Second
-		localIndex  = 0
-
-		// index of the cluster on which a remote namespace with the same name already exists.
-		remoteIndex = 2
-	)
 
 	Context(fmt.Sprintf("Create a namespace inside the cluster '%d' and check what happen if a remote namespaace in the cluster "+
 		"'%d' already exists. Remove the label and check the deletion of the remote namespaces.", localIndex, remoteIndex), func() {
@@ -74,19 +74,19 @@ var _ = Describe("Liqo E2E", func() {
 			By(fmt.Sprintf(" 1 - Creating the remote namespace inside the cluster '%d'", remoteIndex))
 			Eventually(func() error {
 				_, err := util.EnforceNamespace(ctx, testContext.Clusters[remoteIndex].NativeClient,
-					testContext.Clusters[remoteIndex].Cluster, testNamespaceName)
+					testContext.Clusters[remoteIndex].Cluster, namespaceName)
 				return err
 			}, timeout, interval).Should(BeNil())
 
 			By(fmt.Sprintf(" 2 - Creating the local namespace inside the cluster '%d'", localIndex))
 			Eventually(func() error {
 				if _, err := util.EnforceNamespace(ctx, testContext.Clusters[localIndex].NativeClient,
-					testContext.Clusters[localIndex].Cluster, testNamespaceName); err != nil {
+					testContext.Clusters[localIndex].Cluster, namespaceName); err != nil {
 					return err
 				}
 
 				// Do not use liqoctl to create the resource, since it will fail waiting for offloading to complete.
-				return util.CreateNamespaceOffloading(ctx, testContext.Clusters[localIndex].ControllerClient, testNamespaceName,
+				return util.CreateNamespaceOffloading(ctx, testContext.Clusters[localIndex].ControllerClient, namespaceName,
 					offloadingv1alpha1.EnforceSameNameMappingStrategyType, offloadingv1alpha1.LocalAndRemotePodOffloadingStrategyType)
 
 			}, timeout, interval).Should(BeNil())
@@ -94,7 +94,7 @@ var _ = Describe("Liqo E2E", func() {
 			By(" 3 - Getting the NamespaceOffloading resource")
 			Eventually(func() error {
 				if err := testContext.Clusters[localIndex].ControllerClient.Get(ctx,
-					types.NamespacedName{Namespace: testNamespaceName, Name: liqoconst.DefaultNamespaceOffloadingName}, namespaceOffloading); err != nil {
+					types.NamespacedName{Namespace: namespaceName, Name: liqoconst.DefaultNamespaceOffloadingName}, namespaceOffloading); err != nil {
 					return err
 				}
 				if namespaceOffloading.Status.OffloadingPhase != offloadingv1alpha1.SomeFailedOffloadingPhaseType {
@@ -109,7 +109,7 @@ var _ = Describe("Liqo E2E", func() {
 			By(fmt.Sprintf(" 4 - Deleting the remote namespace inside the cluster '%d'", remoteIndex))
 			Eventually(func() error {
 				if err := testContext.Clusters[remoteIndex].ControllerClient.Get(ctx,
-					types.NamespacedName{Name: testNamespaceName}, namespace); err != nil {
+					types.NamespacedName{Name: namespaceName}, namespace); err != nil {
 					return err
 				}
 				return testContext.Clusters[remoteIndex].ControllerClient.Delete(ctx, namespace)
@@ -119,7 +119,7 @@ var _ = Describe("Liqo E2E", func() {
 				"annotation has been created inside the cluster '%d' by the NamespaceMap controller", remoteIndex))
 			Eventually(func() error {
 				if err := testContext.Clusters[remoteIndex].ControllerClient.Get(ctx,
-					types.NamespacedName{Name: testNamespaceName}, namespace); err != nil {
+					types.NamespacedName{Name: namespaceName}, namespace); err != nil {
 					return err
 				}
 				value, ok := namespace.Annotations[liqoconst.RemoteNamespaceManagedByAnnotationKey]
@@ -137,7 +137,7 @@ var _ = Describe("Liqo E2E", func() {
 			namespaceOffloading := &offloadingv1alpha1.NamespaceOffloading{}
 			Eventually(func() metav1.StatusReason {
 				err := testContext.Clusters[localIndex].ControllerClient.Get(ctx,
-					types.NamespacedName{Name: liqoconst.DefaultNamespaceOffloadingName, Namespace: testNamespaceName},
+					types.NamespacedName{Name: liqoconst.DefaultNamespaceOffloadingName, Namespace: namespaceName},
 					namespaceOffloading)
 				_ = testContext.Clusters[localIndex].ControllerClient.Delete(ctx, namespaceOffloading)
 				return apierrors.ReasonForError(err)
@@ -152,9 +152,17 @@ var _ = Describe("Liqo E2E", func() {
 				}
 				Eventually(func() metav1.StatusReason {
 					return apierrors.ReasonForError(testContext.Clusters[i].ControllerClient.Get(ctx,
-						types.NamespacedName{Name: testNamespaceName}, namespace))
+						types.NamespacedName{Name: namespaceName}, namespace))
 				}, timeout, interval).Should(Equal(metav1.StatusReasonNotFound))
 			}
 		})
 	})
+})
+
+var _ = AfterSuite(func() {
+	for i := range testContext.Clusters {
+		Eventually(func() error {
+			return util.EnsureNamespaceDeletion(ctx, testContext.Clusters[i].NativeClient, namespaceName)
+		}, timeout, interval).Should(Succeed())
+	}
 })
